Add travis support for PR/CI testing of JLang
Adds a travis config to build JLang on both MacOS and linux, and run the tests on linux. Due to the new licensing of the oracle JDK7, the linux distro used is Ubuntu precise, which on travis still has the Oracle JDK7 pre-installed. MacOS does not, so running the tests does not work as expected. The travis file installs and/or builds the necessary components, and then builds JLang using its makefiles. The current LLVM jar in the repo does not support Ubuntu 12.04, so LLVM was rebuilt and packaged using the javacpp-presets project, and the jlangc driver was modified to support additional classpath options for running the compiler. The makefiles to build the project and run tests were then modified with the additional classpath only under the travis environment to maintain compatibility with currently working environments.